Restructure project
This commit aims to make the project more accessible to users
